+++ src/index.js
@@ -8,13 +8,15 @@ class ResizeObserverComponent extends Component {
     super(props);
 
     this.state = {
-      width: 0,
-      height: 0
+      width: 1,
+      height: 1
     };
+
+    this.ref = React.createRef();
   }
 
   observeIfNeeded() {
-    const element = this.props.element || ReactDOM.findDOMNode(this);
+    const element = this.ref.current;
 
     if (element && this.element !== element) {
       // clean up after a previous element
@@ -60,24 +62,18 @@ class ResizeObserverComponent extends Component {
   }
 
   componentWillUnmount() {
-    if (this.element) {
-      this.unobserve(this.element);
+    if (this.ref.current) {
+      this.unobserve(this.ref.current);
     }
   }
 
   render() {
-    const size = {
-      width: this.state.width,
-      height: this.state.height
-    };
-
-    return this.props.children(size);
+    return this.props.children(this.ref, this.state.width, this.state.height);
   }
 }
 
 ResizeObserverComponent.propTypes = {
-  children: PropTypes.func.isRequired,
-  element: PropTypes.object
+  children: PropTypes.func.isRequired
 };
 
 export default ResizeObserverComponent;